rest api, swagger

+++ spring_basics/src/main/java/idk/service/notes/NoteServiceImpl.java
@@ -27,7 +27,7 @@ public class NoteServiceImpl implements NoteService {
 
     @Override
     public Note add(Note note) {
-        note.setId(Math.abs(secureRandom.nextLong()));
+        note.setId((long) Math.abs(secureRandom.nextInt()));
         return noteRepository.save(note);
     }
